+++ package.json
@@ -28,7 +28,8 @@
     "@mdx-js/react": "^3.0.0",
     "@popperjs/core": "^2.11.8",
     "@radix-ui/react-avatar": "^1.1.7",
-    "@radix-ui/react-slot": "^1.2.0",
+    "@radix-ui/react-collapsible": "^1.1.12",
+    "@radix-ui/react-slot": "^1.2.3",
     "@tsparticles/react": "^3.0.0",
     "@tsparticles/slim": "^3.8.1",
     "@vercel/analytics": "^1.5.0",
