+++ src/pages/interview-prep/data/behavioralQuestions.ts
@@ -1,0 +1,35 @@
+export const behavioralQuestions = [
+  {
+    category: "Leadership",
+    questions: [
+      "Tell me about a time you led a team through a difficult project",
+      "Describe a situation where you had to influence others without authority",
+      "How do you handle team conflicts?",
+    ],
+  },
+  {
+    category: "Problem Solving",
+    questions: [
+      "Describe the most challenging technical problem you've solved",
+      "Tell me about a time you failed and what you learned",
+      "How do you approach debugging complex issues?",
+    ],
+  },
+  {
+    category: "Communication",
+    questions: [
+      "Explain a complex technical concept to a non-technical person",
+      "Tell me about a time you had to give difficult feedback",
+      "How do you handle disagreements with stakeholders?",
+    ],
+  },
+  {
+    category: "Growth & Learning",
+    questions: [
+      "How do you stay updated with new technologies?",
+      "Tell me about a time you had to learn something completely new",
+      "What's the most important thing you've learned in your career?",
+    ],
+  },
+]
+
