+++ src/pages/interview-prep/data/mockInterviewQuestions.ts
@@ -1,0 +1,20 @@
+export const mockInterviewQuestions = [
+  {
+    type: "technical",
+    question: "Implement a function to reverse a linked list",
+    difficulty: "Medium",
+    hints: ["Think about iterative vs recursive approach", "Consider edge cases like empty list"],
+  },
+  {
+    type: "behavioral",
+    question: "Tell me about a time you had to work with a difficult team member",
+    framework: "STAR Method",
+    hints: ["Situation", "Task", "Action", "Result"],
+  },
+  {
+    type: "system-design",
+    question: "Design a URL shortener like bit.ly",
+    difficulty: "Hard",
+    hints: ["Consider scale", "Database design", "Caching strategy"],
+  },
+]
