+++ src/pages/interview-prep/data/practicePlatforms.ts
@@ -1,0 +1,26 @@
+export const practicePlatforms = [
+  {
+    name: "LeetCode",
+    description: "Most popular coding interview platform",
+    problems: "2000+",
+    difficulty: ["Easy", "Medium", "Hard"],
+    url: "https://leetcode.com/",
+    features: ["Mock Interviews", "Company Tags", "Discussion Forums", "Premium Content"],
+  },
+  {
+    name: "HackerRank",
+    description: "Comprehensive coding challenges and assessments",
+    problems: "1000+",
+    difficulty: ["Easy", "Medium", "Hard"],
+    url: "https://hackerrank.com/",
+    features: ["Skill Assessments", "Certification", "Company Challenges", "Interview Prep"],
+  },
+  {
+    name: "CodeChef",
+    description: "Competitive programming and contests platform",
+    problems: "3000+",
+    difficulty: ["Easy", "Medium", "Hard"],
+    url: "https://www.codechef.com/",
+    features: ["Monthly Contests", "Long Challenge", "Cook-Off", "Practice Problems"],
+  },
+];
